+++ app/src/main/AndroidManifest.xml
@@ -20,19 +20,19 @@
                 <category android:name="android.intent.category.LAUNCHER" />
             </intent-filter>
         </activity>
-        <activity android:name=".activities.MainActivity" />
+        <activity android:name=".activities.main.MainActivity"></activity>
         <meta-data
             android:name="com.google.android.geo.API_KEY"
             android:value="@string/google_maps_key" />
         <activity
-            android:name=".activities.MapsActivity"
+            android:name=".activities.ui.MapsActivity"
             android:label="@string/title_activity_maps" />
-        <activity android:name=".activities.CategoryActivity" />
+        <activity android:name=".activities.category.CategoryActivity" />
         <activity
-            android:name=".activities.EntryActivity"
+            android:name=".activities.entry.EntryActivity"
             android:label="Add Entry" />
         <activity
-            android:name=".activities.VaultActivity"
+            android:name=".activities.vault.VaultActivity"
             android:label="Add vault" />
     </application>
