+++ app/src/main/java/com/aquatic/lucre/activities/auth/LoginActivity.kt
@@ -6,7 +6,7 @@ import androidx.activity.viewModels
 import androidx.appcompat.app.AppCompatActivity
 import androidx.lifecycle.Observer
 import com.aquatic.lucre.R
-import com.aquatic.lucre.activities.MainActivity
+import com.aquatic.lucre.activities.main.MainActivity
 import com.aquatic.lucre.extensions.validate
 import com.aquatic.lucre.viewmodels.UserViewModel
 import kotlinx.android.synthetic.main.activity_login.*
