+++ app/src/main/java/com/aquatic/lucre/activities/dialogs/VaultListDialogFragment.kt
@@ -1,4 +1,4 @@
-package com.aquatic.lucre.activities
+package com.aquatic.lucre.activities.dialogs
 
 import android.content.Intent
 import android.os.Bundle
@@ -7,8 +7,9 @@ import android.view.View
 import android.view.ViewGroup
 import androidx.recyclerview.widget.LinearLayoutManager
 import com.aquatic.lucre.R
-import com.aquatic.lucre.adapters.AdapterListener
+import com.aquatic.lucre.activities.vault.VaultActivity
 import com.aquatic.lucre.adapters.BottomSheetAdapter
+import com.aquatic.lucre.core.AdapterListener
 import com.aquatic.lucre.models.Vault
 import com.google.android.material.bottomsheet.BottomSheetDialogFragment
 import kotlinx.android.synthetic.main.fragment_vault_list_dialog_list_dialog.*
