+++ app/src/main/java/com/aquatic/lucre/activities/ui/ChartFragment.kt
@@ -1,4 +1,4 @@
-package com.aquatic.lucre.activities
+package com.aquatic.lucre.activities.ui
 
 import android.graphics.Color
 import android.os.Bundle
@@ -7,7 +7,6 @@ import android.view.View
 import android.view.ViewGroup
 import androidx.fragment.app.Fragment
 import androidx.fragment.app.activityViewModels
-import androidx.fragment.app.replace
 import androidx.lifecycle.Observer
 import com.aquatic.lucre.R
 import com.aquatic.lucre.viewmodels.CategoryViewModel
