+++ app/src/main/java/com/aquatic/lucre/activities/ui/SpinnerActivity.kt
@@ -1,4 +1,4 @@
-package com.aquatic.lucre.activities
+package com.aquatic.lucre.activities.ui
 
 import android.R
 import android.content.Context
