+++ app/src/main/java/com/aquatic/lucre/adapters/ChartCategoryAdapter.kt
@@ -5,6 +5,8 @@ import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
 import com.aquatic.lucre.R
+import com.aquatic.lucre.core.AdapterListener
+import com.aquatic.lucre.core.BaseAdapter
 import com.aquatic.lucre.models.Category
 import kotlinx.android.synthetic.main.fragment_chart_category_item.view.*
