+++ app/src/main/java/com/aquatic/lucre/core/AdapterListener.kt
@@ -1,4 +1,4 @@
-package com.aquatic.lucre.adapters
+package com.aquatic.lucre.core
 
 interface AdapterListener<T> {
     fun onItemClick(item: T)
