+++ app/src/main/res/layout/activity_category.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".activities.CategoryActivity">
+    tools:context=".activities.category.CategoryActivity">
 
     <RelativeLayout
         android:layout_width="match_parent"
