+++ app/src/main/res/layout/activity_vault.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context="com.aquatic.lucre.activities.VaultActivity">
+    tools:context="com.aquatic.lucre.activities.vault.VaultActivity">
 
     <RelativeLayout
         android:layout_width="match_parent"
