+++ app/src/main/res/layout/fragment_category_list.xml
@@ -5,7 +5,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context="com.aquatic.lucre.activities.fragments.CategoryListFragment">
+    tools:context="com.aquatic.lucre.activities.category.CategoryListFragment">
 
     <androidx.recyclerview.widget.RecyclerView
         android:id="@+id/categoryRecyclerView"
