+++ app/src/main/res/layout/fragment_chart.xml
@@ -3,7 +3,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".activities.ChartFragment">
+    tools:context=".activities.ui.ChartFragment">
 
         <LinearLayout
                 android:layout_width="match_parent"
