+++ app/src/main/res/layout/fragment_dashboard.xml
@@ -3,7 +3,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".activities.DashboardFragment">
+    tools:context=".activities.main.DashboardFragment">
 
     <RelativeLayout
         android:layout_width="match_parent"
