+++ app/src/main/res/layout/fragment_vault_list.xml
@@ -5,7 +5,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context="com.aquatic.lucre.activities.fragments.VaultListFragment">
+    tools:context="com.aquatic.lucre.activities.vault.VaultListFragment">
 
     <FrameLayout
         android:id="@+id/childFragmentContainer"
